# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: The Ministry of Herbology API
  version: 1.10.0
  description: |
    Contract for the Ministry of Herbology. Frozen — only the maintainers
    changes this file, via a recorded decision. Every path declares the parts of the project that
    implements it with `x-parts of the project`; `tests/contract/` enforces that the
    running API matches this document.

    Units: millimetres, °C, litres, lux. Timestamps are RFC 3339 UTC.
servers:
  - url: /api/v1
...
paths:
...
  /almanac/forecast:
    get:
      tags: [almanac]
      summary: 1-day (hourly) and 10-day (daily) forecast
      operationId: getForecast
      parameters:
        - { name: site_id, in: query, required: true, schema: { type: string, format: uuid } }
        - name: horizon
          in: query
          schema: { type: string, enum: [hourly, daily], default: daily }
      responses:
        '200':
          description: Forecast series
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ForecastPoint' }
  /almanac/history:
    get:
      tags: [almanac]
      summary: 1 / 7 / 30-day indoor and outdoor conditions history
      operationId: getHistory
      parameters:
        - name: window
          in: query
          required: true
          schema: { type: string, enum: ['1d', '7d', '30d'] }
        - { name: site_id,     in: query, schema: { type: string, format: uuid } }
        - { name: location_id, in: query, schema: { type: string, format: uuid } }
        - { name: specimen_id, in: query, schema: { type: string, format: uuid } }
        - name: metric
          in: query
          schema:
            type: string
            enum: [temperature_c, humidity_pct, soil_moisture_pct, precip_mm, et0_mm]
      responses:
        '200':
          description: Bucketed series ready for uPlot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Series' }
  /almanac/water-balance/{specimen_id}:
    parameters:
      - { name: specimen_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [almanac]
      summary: Soil water deficit history and current state
      operationId: getWaterBalance
      responses:
        '200':
          description: Water balance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WaterBalance' }
...
  /auth/session:
    get:
      tags: [auth]
      summary: Whether this device is signed in, and whether sign-in is required
      operationId: getAuthSession
      security: []
      responses:
        '200':
          description: The sign-in state of this request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthStatus' }
    post:
      tags: [auth]
      summary: Sign this device in with the household passphrase
      description: >-
        On success sets the `moh_session` cookie (HttpOnly, Secure, SameSite=Lax,
        one year) and answers 204. The passphrase is compared in constant time
        and never logged. Five failures in fifteen minutes from one client, or
        twenty from any mix, answer 429 with `Retry-After`.
      operationId: createAuthSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthSignIn' }
      responses:
        '204': { description: Signed in; the session cookie is set }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429':
          description: Too many failed attempts; try again after `Retry-After` seconds
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [auth]
      summary: Sign this device out
      operationId: deleteAuthSession
      responses:
        '204': { description: Signed out; the cookie is cleared }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/sessions:
    get:
      tags: [auth]
      summary: The household's signed-in devices
      description: Labels and times only — never a token or a hash of one.
      operationId: listAuthSessions
      responses:
        '200':
          description: Active sessions, newest first
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AuthSessionInfo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [auth]
      summary: Sign out every device, this one included
      operationId: deleteAllAuthSessions
      responses:
        '204': { description: Every session revoked; this device's cookie is cleared }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: moh_session
    # the stack's own services, GET only. Nginx deletes this header
    # from every inbound request, so it is unusable from outside the stack.
    service:
      type: apiKey
      in: header
      name: X-Moh-Service-Token
    # the Lunette's device token, valid on /hub/lunette only.
    lunette:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Not signed in, or the session has expired or been revoked
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
        code:   { type: string }
    Confidence:
      type: string
      enum: [high, medium, low, unknown]
    AuthSignIn:
      type: object
      required: [passphrase]
      properties:
        passphrase: { type: string, minLength: 1, maxLength: 1024, writeOnly: true }
    AuthStatus:
      type: object
      description: >-
        `mode: open` is the demo with no passphrase configured — anyone can use
        it, and the UI says so. A deployment is always `required`.
      required: [mode, signed_in, expires_at]
      properties:
        mode:       { type: string, enum: [required, open] }
        signed_in:  { type: boolean }
        expires_at: { type: [string, 'null'], format: date-time }
    AuthSessionInfo:
      type: object
      required: [label, created_at, last_seen_at, current]
      properties:
        label:        { type: string, description: 'Coarse device description, e.g. "Safari on iPhone"' }
        created_at:   { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        current:      { type: boolean, description: This is the device asking }
    Degradation:
      type: object
      description: >-
        One named reason an answer is worth less than a measurement. `caps_at`
        is a ceiling, not a subtraction: two doubts about the same input do not
        make it twice as doubtful, and neither may raise the other's ceiling.
        `detail` is a plain sentence, written to be shown to the reader as-is.
      required: [code, detail, caps_at]
      properties:
        code:    { type: string, description: 'Stable machine name, e.g. et0_fallback' }
        detail:  { type: string }
        caps_at: { $ref: '#/components/schemas/Confidence' }
...
    ForecastPoint:
      type: object
      required: [time]
      properties:
        time:            { type: string, format: date-time }
        temperature_c:   { type: [number, 'null'] }
        temp_min_c:      { type: [number, 'null'] }
        temp_max_c:      { type: [number, 'null'] }
        precip_mm:       { type: [number, 'null'] }
        precip_prob_pct: { type: [number, 'null'] }
        et0_mm:          { type: [number, 'null'] }
        wind_kph:        { type: [number, 'null'] }
        condition:       { type: [string, 'null'] }
        sunrise:         { type: [string, 'null'], format: date-time }
        sunset:          { type: [string, 'null'], format: date-time }
    Series:
      type: object
      description: Column-oriented, ready to hand straight to uPlot.
      required: [metric, times, values]
      properties:
        metric: { type: string }
        unit:   { type: string }
        times:  { type: array, items: { type: integer }, description: Unix seconds }
        values: { type: array, items: { type: [number, 'null'] } }
        min:    { type: array, items: { type: [number, 'null'] } }
        max:    { type: array, items: { type: [number, 'null'] } }
        label:  { type: string }
    WaterBalance:
      type: object
      description: >-
        A modelled soil-water deficit and what it is worth. Under the design
        nothing measures the soil, so this engine alone decides whether an
        outdoor plant is watered; a degraded input must leave here looking
        degraded. `confidence`, `degraded` and `degradations` are required for
        that reason — an answer without them is a degraded number wearing a
        clean one's clothes. See the design.
      required:
        [specimen_id, deficit_mm, threshold_mm, is_due, applies,
         confidence, degraded, degradations]
      properties:
        specimen_id:  { type: string, format: uuid }
        deficit_mm:   { type: number }
        capacity_mm:  { type: number }
        threshold_mm: { type: number }
        k_c:          { type: number }
        is_due:       { type: boolean }
        sensor_override_pct: { type: [number, 'null'] }
        cover_factor: { type: number, description: 'Fraction of rain reaching the soil' }
        status:
          type: string
          enum: [due, satisfied, ok]
          description: >-
            `satisfied` is not `ok`: a watering that was owed and that the
            weather settled stays visible rather than vanishing.
        satisfied_by: { type: [string, 'null'], enum: [rain, irrigation, null] }
        applies:
          type: boolean
          description: >-
            False for indoor specimens, which are watered on an interval rule and not by this engine. The deficit is then shown
            for reference only.
        note: { type: string }
        confidence: { $ref: '#/components/schemas/Confidence' }
        degraded:   { type: boolean }
        degradations:
          type: array
          items: { $ref: '#/components/schemas/Degradation' }
        k_c_confidence: { $ref: '#/components/schemas/Confidence' }
        k_c_source:     { type: [string, 'null'] }
        k_c_is_category_default:
          type: boolean
          description: 'the design — an FAO-56 category default, not a measurement'
        days:
          type: array
          items:
            type: object
            properties:
              day:              { type: string, format: date }
              deficit_mm:       { type: number }
              et0_mm:           { type: number, description: 'K_c x ET0 — the demand made on the soil' }
              reference_et0_mm: { type: [number, 'null'] }
              et0_method:
                type: string
                enum: [ingested, hargreaves, unavailable]
                description: >-
                  the design. `unavailable` does not advance the deficit — an
                  unknown day is not a zero-evaporation day.
              precip_mm:        { type: number }
              gross_precip_mm:  { type: number }
              irrigation_mm:    { type: number }
              is_forecast:      { type: boolean }
              status:           { type: string, enum: [due, satisfied, ok] }
              satisfied_by:     { type: [string, 'null'], enum: [rain, irrigation, null] }
...
security:
  - session: []